+++ src/main/java/org/sopt/diary/api/DiaryController.java
@@ -3,12 +3,14 @@ package org.sopt.diary.api;
 import java.util.List;
 import org.sopt.diary.constant.ResponseMessage;
 import org.sopt.diary.dto.Diary;
-import org.sopt.diary.dto.DiaryListResponse;
-import org.sopt.diary.dto.DiaryRequest;
+import org.sopt.diary.dto.request.DiaryRequest;
+import org.sopt.diary.dto.response.DiaryDetailResponse;
+import org.sopt.diary.dto.response.DiaryListResponse;
 import org.sopt.diary.service.DiaryService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -22,7 +24,7 @@ public class DiaryController {
     }
 
     @PostMapping("/diary")
-    ResponseEntity<String> createDiary(@RequestBody DiaryRequest diaryRequest) {
+    ResponseEntity<String> create(@RequestBody DiaryRequest diaryRequest) {
         Diary diary = Diary.toDiaryDto(diaryRequest);
         diaryService.createDiary(diary);
 
@@ -31,9 +33,17 @@ public class DiaryController {
 
     @GetMapping("/diary")
     ResponseEntity<DiaryListResponse> getList() {
-        List<Diary> diaryList = diaryService.getList();
+        List<Diary> diaryList = diaryService.getDiaryList();
         DiaryListResponse diaryListResponse = DiaryListResponse.toDiaryListResponse(diaryList);
 
         return ResponseEntity.ok(diaryListResponse);
+    }
+
+    @GetMapping("/diary/{id}")
+    ResponseEntity<DiaryDetailResponse> getDetail(@PathVariable Long id) {
+        Diary diary = diaryService.getDiaryDetail(id);
+        DiaryDetailResponse diaryDetailResponse = DiaryDetailResponse.toDiaryDetailResponse(diary);
+
+        return ResponseEntity.ok(diaryDetailResponse);
     }
 }
